+++ tests/integration/test_delete_count.py
@@ -15,23 +15,23 @@
 # specific language governing permissions and limitations
 # under the License.
 # pylint:disable=redefined-outer-name
-from datetime import datetime
+import random
+from datetime import datetime, timedelta
 from typing import Generator, List
 
 import pyarrow as pa
 import pytest
+from pyarrow import compute as pc
 from pyspark.sql import SparkSession
 
+from pyiceberg.catalog import Catalog
 from pyiceberg.catalog.rest import RestCatalog
 from pyiceberg.exceptions import NoSuchTableError
-from pyiceberg.expressions import AlwaysTrue, EqualTo
-from pyiceberg.manifest import ManifestEntryStatus
-from pyiceberg.partitioning import PartitionField, PartitionSpec
+from pyiceberg.expressions import And, EqualTo, GreaterThanOrEqual, LessThan
 from pyiceberg.schema import Schema
 from pyiceberg.table import Table
-from pyiceberg.table.snapshots import Operation, Summary
-from pyiceberg.transforms import IdentityTransform
-from pyiceberg.types import FloatType, IntegerType, LongType, NestedField, StringType, TimestampType
+from pyiceberg.transforms import HourTransform, IdentityTransform
+from pyiceberg.types import LongType, NestedField, StringType
 
 
 def run_spark_commands(spark: SparkSession, sqls: List[str]) -> None:
@@ -67,8 +67,8 @@ def test_partitioned_table_delete_full_file(spark: SparkSession, session_catalog
             f"DROP TABLE IF EXISTS {identifier}",
             f"""
             CREATE TABLE {identifier} (
-                number_partitioned  int,
-                number              int
+                number_partitioned  long,
+                number              long
             )
             USING iceberg
             PARTITIONED BY (number_partitioned)
@@ -91,99 +91,69 @@ def test_partitioned_table_delete_full_file(spark: SparkSession, session_catalog
     assert tbl.scan().to_arrow().to_pydict() == {"number_partitioned": [11, 11], "number": [20, 30]}
 
     assert tbl.scan().count() == len(tbl.scan().to_arrow())
+    filter = And(EqualTo("number_partitioned", 11), GreaterThanOrEqual("number", 5))
+    assert tbl.scan(filter).count() == len(tbl.scan(filter).to_arrow())
+    N = 10
+    d = {
+        "number_partitioned": pa.array([i * 10 for i in range(N)]),
+        "number": pa.array([random.choice([10, 20, 40]) for _ in range(N)]),
+    }
+    with tbl.update_spec() as update:
+        update.add_field("number", transform=IdentityTransform())
+
+    data = pa.Table.from_pydict(d)
+
+    tbl.overwrite(df=data, overwrite_filter=filter)
 
 
 @pytest.mark.integration
-@pytest.mark.filterwarnings("ignore:Merge on read is not yet supported, falling back to copy-on-write")
-def test_delete_partitioned_table_positional_deletes(spark: SparkSession, session_catalog: RestCatalog) -> None:
-    identifier = "default.table_partitioned_delete"
+def test_rewrite_manifest_after_partition_evolution(session_catalog: Catalog) -> None:
+    random.seed(876)
+    N = 1440
+    d = {
+        "timestamp": pa.array([datetime(2023, 1, 1, 0, 0, 0) + timedelta(minutes=i) for i in range(N)]),
+        "category": pa.array([random.choice(["A", "B", "C"]) for _ in range(N)]),
+        "value": pa.array([random.gauss(0, 1) for _ in range(N)]),
+    }
+    data = pa.Table.from_pydict(d)
 
-    run_spark_commands(
-        spark,
-        [
-            f"DROP TABLE IF EXISTS {identifier}",
-            f"""
-            CREATE TABLE {identifier} (
-                number_partitioned  int,
-                number              int
-            )
-            USING iceberg
-            PARTITIONED BY (number_partitioned)
-            TBLPROPERTIES(
-                'format-version' = 2,
-                'write.delete.mode'='merge-on-read',
-                'write.update.mode'='merge-on-read',
-                'write.merge.mode'='merge-on-read'
-            )
-        """,
-            f"""
-            INSERT INTO {identifier} VALUES (10, 20), (10, 30), (10, 40)
-        """,
-            # Generate a positional delete
-            f"""
-            DELETE FROM {identifier} WHERE number = 30
-        """,
-        ],
+    try:
+        session_catalog.drop_table(
+            identifier="default.test_error_table",
+        )
+    except NoSuchTableError:
+        pass
+
+    table = session_catalog.create_table(
+        "default.test_error_table",
+        schema=data.schema,
+    )
+    with table.update_spec() as update:
+        update.add_field("timestamp", transform=HourTransform())
+
+    table.append(data)
+    assert table.scan().count() == len(table.scan().to_arrow())
+
+    with table.update_spec() as update:
+        update.add_field("category", transform=IdentityTransform())
+
+    data_ = data.filter(
+        (pc.field("category") == "A")
+        & (pc.field("timestamp") >= datetime(2023, 1, 1, 0))
+        & (pc.field("timestamp") < datetime(2023, 1, 1, 1))
     )
 
-    tbl = session_catalog.load_table(identifier)
-
-    assert tbl.scan().count() == len(tbl.scan().to_arrow())
-
-    # Will rewrite a data file without the positional delete
-    tbl.delete(EqualTo("number", 40))
-
-    assert [snapshot.summary.operation.value for snapshot in tbl.snapshots()] == ["append", "overwrite", "overwrite"]
-    assert tbl.scan().to_arrow().to_pydict() == {"number_partitioned": [10], "number": [20]}
-
-    assert tbl.scan().count() == len(tbl.scan().to_arrow())
-
-    run_spark_commands(
-        spark,
-        [
-            f"""
-            INSERT INTO {identifier} VALUES (10, 20), (10, 30), (10, 40), (20, 30)
-        """,
-            # Generate a positional delete
-            f"""
-            DELETE FROM {identifier} WHERE number = 30
-        """,
-        ],
+    filter = And(
+        And(
+            GreaterThanOrEqual("timestamp", datetime(2023, 1, 1, 0).isoformat()),
+            LessThan("timestamp", datetime(2023, 1, 1, 1).isoformat()),
+        ),
+        EqualTo("category", "A"),
     )
-
-    tbl = session_catalog.load_table(identifier)
-
-    assert tbl.scan().count() == len(tbl.scan().to_arrow())
-
-
-    run_spark_commands(
-        spark,
-        [
-            # Generate a positional delete
-            f"""
-            DELETE FROM {identifier} WHERE number = 30
-        """,
-            f"""
-            INSERT INTO {identifier} VALUES (10, 20), (10, 30), (10, 40), (20, 30)
-        """,
-            # Generate a positional delete
-            f"""
-                DELETE FROM {identifier} WHERE number = 20
-            """,
-        ],
-    )
-
-    tbl = session_catalog.load_table(identifier)
-
-    assert tbl.scan().count() == len(tbl.scan().to_arrow())
-
-
-    filter_on_partition = "number_partitioned = 10"
-    scan_on_partition = tbl.scan(row_filter=filter_on_partition)
-    assert  scan_on_partition.count() == len(scan_on_partition.to_arrow())
-
-
-    filter = "number = 10"
-    scan = tbl.scan(row_filter=filter)
-    assert  scan.count() == len(scan.to_arrow())
-
+    # filter = GreaterThanOrEqual("timestamp", datetime(2023, 1, 1, 0).isoformat())
+    # filter = LessThan("timestamp", datetime(2023, 1, 1, 1).isoformat())
+    # filter = EqualTo("category", "A")
+    # assert table.scan().plan_files()[0].file.partition == {"category": "A"}
+    assert table.scan().count() == len(table.scan().to_arrow())
+    assert table.scan(filter).count() == len(table.scan(filter).to_arrow())
+    table.overwrite(df=data_, overwrite_filter=filter)
